refactore addLayer
