+++ elements/layercontrol/src/components/addLayers.js
@@ -1,16 +1,15 @@
 import { LitElement, html, nothing } from "lit";
 import "./layer";
 import "./layerGroup";
+import { isMapUrlValid, isLayerJSONValid } from "../helpers";
 import {
-  handleAddLayerMethod,
-  handleJsonInputChangeMethod,
-  handleOpenCloseTabMethod,
-  handleUrlInputChangeMethod,
-  handleUrlLayerMethod,
-  handleWMSSearchURLMethod,
-  isLayerJSONValid,
-  isMapUrlValid,
-} from "../helpers";
+  addLayerMethod,
+  urlLayerMethod,
+  openCloseTabMethod,
+  wmsSearchURLMethod,
+  urlInputChangeMethod,
+  jsonInputChangeMethod,
+} from "../methods/add-layer";
 
 /**
  * EOxLayerControlAddLayers is a class that extends LitElement and provides functionalities
@@ -83,20 +82,20 @@ export class EOxLayerControlAddLayers extends LitElement {
   }
 
   /**
-   * Handles changes in the input field by invoking the 'handleUrlInputChangeMethod'.
+   * Handles changes in the input field by invoking the 'urlInputChangeMethod'.
    * This method is triggered upon an @input change event.
    *
    * @param {Event} evt - The input change event triggering the method.
    */
   #handleURLChange(evt) {
-    handleUrlInputChangeMethod(evt, this);
+    urlInputChangeMethod(evt, this);
   }
 
   /**
    * Asynchronously handles WMS URL search, invokes further processing if data is available.
    */
   async #handleWMSSearchURL() {
-    const data = await handleWMSSearchURLMethod(this);
+    const data = await wmsSearchURLMethod(this);
     if (data) this.#handleUrlLayerMethod(data);
   }
 
@@ -106,33 +105,33 @@ export class EOxLayerControlAddLayers extends LitElement {
    * @param {{"Name": string}} layer - The layer information triggering the method.
    */
   #handleUrlLayerMethod(layer) {
-    handleUrlLayerMethod(layer, this);
-    handleAddLayerMethod(this);
+    urlLayerMethod(layer, this);
+    addLayerMethod(this);
   }
 
   /**
    * Initiates the addition of layers, triggering the handleAddLayerMethod.
    */
   #handleAddLayer() {
-    handleAddLayerMethod(this);
+    addLayerMethod(this);
   }
 
   /**
    * Handles input field changes by invoking the 'handleJsonInputChangeMethod'.
    *
-   * @param {Event} evt - The input change event triggering the method.
+   * @param {{target: { value: string }}} evt - The input change event triggering the method.
    */
   #handleInputChange(evt) {
-    handleJsonInputChangeMethod(evt, this);
+    jsonInputChangeMethod(evt, this);
   }
 
   /**
-   * Handles tab changes by invoking the 'handleOpenCloseTabMethod'.
+   * Handles tab changes by invoking the 'openCloseTabMethod'.
    *
    * @param {string} tab - The tab identifier triggering the method.
    */
   #handleOpenCloseTab(tab) {
-    handleOpenCloseTabMethod(tab, this);
+    openCloseTabMethod(tab, this);
   }
 
   /**
@@ -206,7 +205,9 @@ export class EOxLayerControlAddLayers extends LitElement {
                           <!-- Button to add layer -->
                           <button
                             class="add-layer-icon icon"
-                            @click=${() => this.#handleUrlLayerMethod(layer)}
+                            @click=${() =>
+                              //@ts-ignore
+                              this.#handleUrlLayerMethod(layer)}
                           ></button>
                         </li>
                       `
