+++ elements/layercontrol/src/methods/add-layer/index.js
@@ -1,0 +1,6 @@
+export { default as jsonInputChangeMethod } from "./json-input-change";
+export { default as addLayerMethod } from "./add-layer";
+export { default as urlInputChangeMethod } from "./url-input-change";
+export { default as wmsSearchURLMethod } from "./wms-search-url";
+export { default as urlLayerMethod } from "./url-layer";
+export { default as openCloseTabMethod } from "./open-close-tab";
